requirements.txt
krakenex==2.1.0
toml==0.10.2
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0

